Added recursive function for calculating midpoints, values stored in array

+++ ProceduralModeler/ProceduralModeler/Terrain.h
@@ -4,5 +4,5 @@ class Terrain {
 public:
 
 	float generateRandomOffset(float aX, float aY, float bX, float bY, float s);
-	float generateMidpoint(float a, float b, float r);
+	int generateMidpoint(float a, float b, float r);
 };
